+++ Printf/srcs/basic/ft_putuint.c
@@ -6,16 +6,16 @@
 /*   By: juspende <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/04 15:06:32 by juspende          #+#    #+#             */
-/*   Updated: 2017/11/22 17:02:46 by juspende         ###   ########.fr       */
+/*   Updated: 2017/11/28 11:44:14 by juspende         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "../../include/ft_printf.h"
 
-void	ft_putuint(unsigned int nb, t_flag *flag)
+void	ft_putuint(uintmax_t nb, t_flag *flag)
 {
-	unsigned int	i;
+	uintmax_t	i;
 
 	i = nb;
 	if (i > 9)
